// app/types/interfaces.ts
export interface Ath_response {
    phone: string;
    name: string;
    email: string;
}

export interface employeeProp {
    id: string;
    photo: string;
    name: string;
    email: string;
    phone: string
    BI: string;
    address: string;
    fatherName: string;
    matherName: string;
    altura: string;
    bron: string;
    BI_epired: string;
    stade: string;
    cargo: string;
    Start: string;
    job: string;
    AgenteNumber: string;
}


export interface cloudinaryProp {
    asset_id: string;
    bytes: number;
    created_at: string;
    display_name: string;
    format: string;
    original_filename: string;
    resource_type: string;
    secure_url: string;
    signature: string;
}